整体修改1
整体修改1-从应答日志

+++ src/com/jlj/util/ExportExcelForAlarmRecord.java
@@ -54,10 +54,10 @@ public class ExportExcelForAlarmRecord {
      * @param cells
      * @param rowIndex
      */
-    public static void createTableRow(List<String> cells,short rowIndex)
+    public static void createTableRow(List<String> cells,int rowIndex)
     {
         //创建第rowIndex行
-        HSSFRow row = demoSheet.createRow((short) rowIndex);
+        HSSFRow row = demoSheet.createRow(rowIndex);
         for(short i = 0;i < cells.size();i++)
         {
             //创建第i个单元格
@@ -117,7 +117,7 @@ public class ExportExcelForAlarmRecord {
     			list.add(alarmrecord.getPhones());
     			list.add(alarmrecord.getSendreturn());
     			
-    			createTableRow(list,(short)(k+1));
+    			createTableRow(list,(k+1));
     		}
 		} catch (Exception e) {
 			e.printStackTrace();
